+++ testfiles/test_python.py
@@ -1,3 +1,3 @@
-word = str(input()) 
+word = str(input())
 print("Hello, " + word)
 print("your are awsm!")